refactor; implement Notebook page loading before receiving the list of experts
