+++ hwproj.front/src/components/Experts/InviteModal.tsx
@@ -6,12 +6,12 @@ import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import ApiSingleton from "../api/ApiSingleton";
+import ApiSingleton from "../../api/ApiSingleton";
 import Typography from "@material-ui/core/Typography";
 import Grid from '@material-ui/core/Grid';
 import {IconButton} from "@material-ui/core";
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import {CoursePreviewView, HomeworkViewModel, CreateCourseFilterViewModel, AccountDataDto} from "../api";
+import {CoursePreviewView, HomeworkViewModel, CreateCourseFilterViewModel, AccountDataDto} from "../../api";
 import {Select, MenuItem, InputLabel, FormControl, Autocomplete} from "@mui/material";
 
 interface IInviteExpertProps {
